Update params, some fixes and save component
