+++ profsea/utils/utils.py
@@ -1,4 +1,6 @@
+import os
 from pathlib import Path
+from typing import Dict
 import zipfile
 
 import dask.array as da
@@ -175,3 +177,78 @@ def fetch_zenodo_fingerprints(
         # 4. Clean up the zip file
         if zip_path.exists():
             zip_path.unlink()
+
+
+def save_components(
+    self,
+    components: Dict[str, xr.DataArray],
+    scenario_name: str,
+    output_prefix: str = "projection",
+    output_dir: str = ".",
+    output_format: str = "zarr",
+) -> None:
+    """
+    Stream all regional sea level projections to disk in a single file/store.
+
+    Parameters
+    ----------
+    components: Dict[str, xr.DataArray]
+        Dictionary of component names and their corresponding Xarray DataArrays.
+    output_format: str
+        Format to save the output in. Must be either 'netcdf' or 'zarr'.
+    output_dir: str
+        Directory to save components to.
+    scenario_name: str
+        Name of the scenario you've run the emulator for.
+    output_prefix: str
+        Prefix for the output file name (e.g., 'projection' will result in 'ssp
+
+    Returns
+    -------
+    None
+    """
+    ds = xr.Dataset(components)
+
+    output_format = output_format.lower()
+    if output_format not in ["netcdf", "zarr"]:
+        raise ValueError("output_format must be either 'netcdf' or 'zarr'.")
+
+    Path(output_dir).mkdir(parents=True, exist_ok=True)
+    encoding = {}
+
+    # Sort out Zarr encoding
+    if output_format == "zarr":
+        import numcodecs
+        from numcodecs.zarr3 import Blosc
+
+        compressor = Blosc(cname="zstd", clevel=5, shuffle=numcodecs.Blosc.BITSHUFFLE)
+
+    # Set the encoding/compression for each variable based on the output format
+    for name, component in components.items():
+        if output_format == "netcdf":
+            encoding[name] = {"zlib": True, "complevel": 1, "dtype": "float32"}
+        elif output_format == "zarr":
+            encoding[name] = {"compressor": compressor, "dtype": "float32"}
+
+    file_name = f"{scenario_name}_{output_prefix}"
+
+    # Stream the computation and write to disk
+    if output_format == "netcdf":
+        out_path = os.path.join(output_dir, f"{file_name}.nc")
+        with console.status(
+            "[bold cyan]Computing and saving NetCDF...[/bold cyan]", spinner="dots"
+        ):
+            ds.compute().to_netcdf(out_path, encoding=encoding)
+        console.log(f"[bold green]✓ Successfully saved NetCDF:[/bold green] {out_path}")
+
+    elif output_format == "zarr":
+        out_path = os.path.join(output_dir, f"{file_name}.zarr")
+        with console.status(
+            "[bold cyan]Streaming computation and saving Zarr...[/bold cyan]",
+            spinner="dots",
+        ):
+            ds.to_zarr(out_path, encoding=encoding, mode="w", compute=True)
+        console.log(f"[bold green]✓ Successfully saved Zarr:[/bold green] {out_path}")
+
+    dims_str = ", ".join(ds[name].dims)
+    console.log(f"Output shape was {ds[name].shape} ({dims_str})")
